Android 16.0.0 release 4

+++ libs/binder/trusty/include_mock/uapi/trusty_uuid.h
@@ -15,6 +15,8 @@
  */
 #pragma once
 
+#include <stdint.h>
+
 typedef struct uuid {
-    int placeholder;
+    uint64_t placeholder;
 } uuid_t;
